requirements.txt
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
deepl==1.15.0
fuzzywuzzy==0.18.0
greenlet==2.0.2
idna==3.4
joblib==1.2.0
Levenshtein==0.21.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.3
pandas==2.0.1
protobuf==3.20.3
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.0
pytz==2023.3
rapidfuzz==3.1.0
regex==2023.6.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.12
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.3
